updated styling of sidebar and topbar
