+++ client/package.json
@@ -17,8 +17,7 @@
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
-    "sass": "^1.54.3",
-    "web-vitals": "^1.1.2"
+    "sass": "^1.54.3"
   },
   "scripts": {
     "start": "PORT=4000 react-scripts start",
