--- client/src/App.css
@@ -1,61 +1,0 @@
-
-.sidenav {
-  display: flex;
-  height: 100%; /* 100% Full-height */
-  width: auto; /* 0 width - change this with JavaScript */
-  position: fixed; /* Stay in place */
-  z-index: 1; /* Stay on top */
-  top: 0; /* Stay at the top */
-  left: 0;
-  background-color: rgb(153, 145, 145); /* Black*/
-  overflow-x: hidden; /* Disable horizontal scroll */
-  padding-top: 60px; /* Place content 60px from the top */
-  transition: 0.5s; /* 0.5 second transition effect to slide in the sidenav */
-}
-
-.sidenavLink {
-  text-decoration: none;
-  font-size: 35px;
-  color: #000000;
-  display: block;
-  transition: 0.3s;
-
-}
-
-ul{
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-}
-
-.main{
-  display: grid;
-}
-
-.topHeader {
-  float: left;
-}
-
-[data-initials]:before {
-  background: #099bdd;
-  color: white;
-  opacity: 1; 
-  content: attr(data-initials); 
-  display: inline-block; 
-  font-weight: bold; 
-  border-radius: 50%; 
-  vertical-align: middle; 
-  margin-right: 0.5em; 
-  width: 50px; 
-  height: 50px; 
-  line-height: 50px; 
-  text-align: center; 
-  }
-
-  input[type="radio"]{
-    display:none;
-  }
-  .star {
-    cursor: pointer;
-    transition: color 200ms;
-  }
