+++ client/src/components/app/app.scss
@@ -1,0 +1,42 @@
+
+
+
+
+
+/* ul{
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+} */
+
+/* .main{
+  display: grid;
+} */
+
+/* .topHeader {
+  float: left;
+} */
+
+[data-initials]:before {
+  background: #099bdd;
+  color: white;
+  opacity: 1; 
+  content: attr(data-initials); 
+  display: inline-block; 
+  font-weight: bold; 
+  border-radius: 50%; 
+  vertical-align: middle; 
+  margin-right: 0.5em; 
+  width: 50px; 
+  height: 50px; 
+  line-height: 50px; 
+  text-align: center; 
+  }
+
+  input[type="radio"]{
+    display:none;
+  }
+  .star {
+    cursor: pointer;
+    transition: color 200ms;
+  }
