+++ client/src/components/navigation/navBar.scss
@@ -1,0 +1,37 @@
+.side-nav{
+    background: #181818;
+    width: 100px;
+    height: 100%;
+    position: absolute;
+    top: 0;
+    z-index: 3;
+    min-height: 500px;
+
+
+    ul{
+        position:absolute;
+        top: 100px;
+        width: 100%;
+        display: block;
+        padding:0;
+        list-style: none;
+        text-align: center;
+        margin: 0;
+
+        li{
+            .sideLink{
+                color: white;
+                padding: 10px 0;
+                display: block;
+                font-size: 20px;
+                line-height: 16px;
+                text-decoration: none;
+
+                &:hover {
+                    color: #ffd700;
+                }
+            }
+        }
+        }
+}
+
