+++ client/src/components/topbar/topbar.scss
@@ -1,0 +1,16 @@
+.topbar{
+    height: 100px;
+    width: 100vw;
+    background-color: #181818;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: center;
+
+    .link{
+        color: white;
+        font-size: 25px;
+        text-decoration: none;
+        padding: 20px 20px;
+    }
+}
