+++ client/src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {BrowserRouter as Router} from 'react-router-dom'
 import './index.css';
-import App from './components/App';
-import reportWebVitals from './reportWebVitals';
+import App from './components/app/App.jsx';
+
 
 ReactDOM.render(
   <Router>
@@ -12,7 +12,4 @@ ReactDOM.render(
   document.getElementById('root')
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+
